# Extract the app-server boot function; rename its binary
## Motivation

`@aikirun/app-server` is the standalone app that boots the Aiki server as a process — it loads config, wires optional Redis, serves over HTTP, and handles graceful shutdown. Today all of that lives in one entry module behind an `import.meta.main` guard, and compiles to a binary named `aiki-server`.

Both problems below sit ahead of a planned `aiki` binary: a single downloadable executable that will combine database migration and serving into the standalone distribution.

- **The boot logic can't be reused.** The `aiki` binary will serve by running the same boot sequence `@aikirun/app-server` already has. But that sequence is trapped inside a run-only entry module — it executes on import and exports nothing, so nothing else can call it.
- **`aiki-server` names two different binaries.** `@aikirun/server` (the SDK library) ships an operator bin called `aiki-server` that applies migrations. `@aikirun/app-server` (this app) compiles to a serve-only binary of the same name. Two unrelated `aiki-server`s.

## Solution

Split `@aikirun/app-server` into three files with distinct jobs:

- **the boot logic** — exported as `startAppServer()`.
- **the entry** — imports and runs `startAppServer()`; this is what compiles to the binary and backs `bun run start`.
- **the barrel** — re-exports `startAppServer`, so `import { startAppServer } from "@aikirun/app-server"` works. The `aiki` binary will import it here.

Rename this app's compiled binary `aiki-server` → `aiki-app-server`, leaving the name `aiki-server` to the SDK's operator bin. The published image (`aikirun/server`) and the compose container (`aiki-server`) keep their names — an operator deals with one server, and those names are unambiguous in their own namespaces. Only the compiled binary, and the Dockerfile `CMD` that runs it, change.
