+++ app/server/src/main.ts
@@ -1,0 +1,3 @@
+import { startAppServer } from "./serve";
+
+await startAppServer();
